Split ExpressionParserConstants to 2 files
