+++ src/app/digital/utils/ExpressionParser/GenerateTokens.ts
@@ -1,5 +1,6 @@
 import { SubEquals } from "core/utils/StringUtils";
-import { FormatLabels, InputToken, OpsArray, Token } from "./Constants";
+import { FormatLabels, InputToken, Token } from "./Constants/DataStructures";
+import { OpsArray } from "./Constants/Objects";
 
 
 /**
